requirements.txt
pyannote.audio
torch
torchaudio
ffmpeg-python
python-dotenv
numpy
scipy
